requirements.txt
aenum==2.2.4
argparse-addons==0.6.0
bitstruct==8.11.0
cantools==35.2.0
diskcache==4.1.0
filelock==3.0.12
mypy-extensions==0.4.3
nose==1.3.7
pymodbus==2.3.0
pyserial==3.4
python-can==4.0.0.dev0
six==1.15.0
textparser==0.23.0
wrapt==1.12.1
